update: separate cli logic
