+++ lib/converter.js
@@ -1,20 +1,15 @@
-#!/usr/bin/env node
-
-var PROGRAM_VERSION    = require('./package.json').version;
-
-var GraphicsConverter  = require('./lib/GraphicsConverter');
-var TransformConverter = require('./lib/TransformConverter');
-var TweenConverter     = require('./lib/TweenConverter');
-var RectangleConverter = require('./lib/RectangleConverter');
-var SetRectangleConverter = require('./lib/SetRectangleConverter');
-var TextConverter      = require('./lib/TextConverter');
-var CacheConverter     = require('./lib/CacheConverter');
+var GraphicsConverter  = require('./GraphicsConverter');
+var TransformConverter = require('./TransformConverter');
+var TweenConverter     = require('./TweenConverter');
+var RectangleConverter = require('./RectangleConverter');
+var SetRectangleConverter = require('./SetRectangleConverter');
+var TextConverter      = require('./TextConverter');
+var CacheConverter     = require('./CacheConverter');
 var fs                 = require('fs');
 var path               = require('path');
-var program            = require('commander');
 var _                  = require('lodash');
 
-function TFCRatioConverter(file, ratio, base) {
+var TFCRatioConverter = function (file, ratio, base) {
     var source = path.resolve(__dirname + '/', file);
 
     if ( ! fs.existsSync(source) ) {
@@ -26,7 +21,7 @@ function TFCRatioConverter(file, ratio, base) {
     this.ratio      = ratio;
     this.base       = base;
     this.line       = 0;
-}
+};
 
 TFCRatioConverter.prototype.convert = function() {
 
@@ -279,46 +274,10 @@ TFCRatioConverter.prototype.convert = function() {
 
 TFCRatioConverter.prototype.getConvertedFileName = function() {
     var fileName  = path.basename(this.sourceFile).split('.'),
-        ratio = _.isNaN(this.ratio * 10) ? this.ratio : this.ratio * 10;
+        ratio = _.isNaN(this.ratio * 10) ? this.ratio : this.ratio * 10,
         extension = fileName.pop();
 
     return process.cwd() + '/' + fileName.join('.') + '_' + ratio + '.' + extension;
 };
 
-
-program
-    .version(PROGRAM_VERSION)
-    .usage('<file> [options]')
-    .option('-o, --output <output>', 'convert output file. If empty, output filename will be replaced [target_file_body]_[ratio*10].js.')
-    .option('-r, --ratio <ratio>', 'convert ratio. path caluculated by this rate. ( default 1 )', 1)
-    .option('-b, --base <base>', 'convert base ratio (default ratio 2)', 2)
-    //.option('-f, --file <file>', 'file [filename]')
-    //.option('-l, --log', 'log mode')
-    .parse(process.argv);
-
-    //console.log(program.args);
-
-var filePath = program.args[0];
-if (!filePath) {
-    program.help();
-    return console.log('should set -f file name!');
-}
-
-if (program.ratio) console.log('  - ratio', program.ratio);
-if (program.base) console.log('  - base', program.base);
-if (program.output) console.log('  - output', program.output);
-console.log('  - convert', filePath);
-
-
-//var tfc    = new TFCRatioConverter(path.resolve(process.cwd() + '/' + targetFile), Argv.get('ratio'), Argv.get('base'));
-var tfc    = new TFCRatioConverter(path.resolve(process.cwd() + '/' + filePath), program.ratio, program.base);
-var buffer = tfc.convert();
-var dest   = program.output || tfc.getConvertedFileName();
-
-fs.writeFile(dest, buffer, {encoding:'utf8'}, function(err) {
-    if ( err ) {
-        return console.log('Oops! failed to write converted source.');
-    }
-
-    console.log('Success! converted file created. ->', dest);
-});
+module.exports = TFCRatioConverter;
